+++ src/packages/piral-core/src/components/ExtensionSlot.tsx
@@ -4,7 +4,12 @@ import { useGlobalState } from '../hooks';
 import { defaultRender } from '../utils';
 import { ExtensionSlotProps } from '../types';
 
-export function ExtensionSlot<T = any>({ name, render = defaultRender, empty, params }: ExtensionSlotProps<T>) {
+export function ExtensionSlot<T extends string>({
+  name,
+  render = defaultRender,
+  empty,
+  params,
+}: ExtensionSlotProps<T>) {
   const extensions = useGlobalState(s => s.registry.extensions[name] || []);
   return render(
     extensions.length === 0 && isfunc(empty)
@@ -20,4 +25,5 @@ export function ExtensionSlot<T = any>({ name, render = defaultRender, empty, pa
         )),
   );
 }
-(ExtensionSlot as React.FC<ExtensionSlotProps>).displayName = `ExtensionSlot`;
+
+ExtensionSlot.displayName = `ExtensionSlot`;