+++ src/packages/piral-core/src/types/custom.ts
@@ -4,6 +4,11 @@
 export interface PiletCustomApi {}
 
 /**
+ * Custom meta data to include for pages.
+ */
+export interface PiralCustomPageMeta {}
+
+/**
  * Custom state extensions defined outside of piral-core.
  */
 export interface PiralCustomState {}
@@ -22,6 +27,11 @@ export interface PiralCustomActions {}
  * Custom events defined outside of piral-core.
  */
 export interface PiralCustomEventMap {}
+
+/**
+ * Custom extension slots outside of piral-core.
+ */
+export interface PiralCustomExtensionSlotMap {}
 
 /**
  * Custom parts of the global registry state defined outside of piral-core.