+++ src/packages/piral-core/src/types/state.ts
@@ -13,6 +13,7 @@ import {
   PageComponentProps,
   ExtensionComponentProps,
   PiletsBag,
+  PiralPageMeta,
 } from './api';
 import {
   ComponentConverters,
@@ -41,10 +42,11 @@ export interface BaseRegistration {
 
 export interface PageRegistration extends BaseRegistration {
   component: WrappedComponent<PageComponentProps>;
+  meta: PiralPageMeta;
 }
 
 export interface ExtensionRegistration extends BaseRegistration {
-  component: WrappedComponent<ExtensionComponentProps<any>>;
+  component: WrappedComponent<ExtensionComponentProps<string>>;
   reference: any;
   defaults: any;
 }