+++ src/packages/piral/polyfills.js
@@ -3,3 +3,4 @@ require('url-polyfill');
 require('whatwg-fetch');
 require('current-script-polyfill');
 require('core-js/stable');
+require('regenerator-runtime/runtime');