+++ src/samples/sample-piral-core/src/pilets/dashboard.tsx
@@ -11,7 +11,7 @@ export const DashboardPilet: Pilet = {
   version: '1.0.0',
   hash: '3',
   setup(piral) {
-    piral.registerExtension<{ children: React.ReactChild }>('dashboard', ({ params: { children } }) => (
+    piral.registerExtension('dashboard', ({ params: { children } }) => (
       <div className="tiles">{children}</div>
     ));
   },